+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2
-*
+* Change initLocales signature to accept base directory
+* Replace opts.dir with opts.localesDir appended to base directory
 
 ### 0.0.1
 * Initial version